+++ index.ts
@@ -1,5 +1,4 @@
 import {cyan, red, yellow} from 'chalk';
-import {EventEmitter} from 'events';
 
 import amqplib = require('amqplib/callback_api');
 import objectAssign = require('object-assign');
@@ -11,15 +10,12 @@ import {stringify, stringifyError, parse, parseError} from './lib/serialization'
 
 const DEFAULT_RPC_EXPIRY = 15000; // 15 seconds
 
-function maybeFromCallback<T>(fn: ((done: Rabbitr.Callback<T>) => void) | (() => PromiseLike<T>)): Bluebird<T> {
-  let callback: Rabbitr.Callback<T>;
-  let promise = Bluebird.fromCallback<T>(_callback => (callback = _callback) && void 0);
-
-  let val = (fn as Function)(callback);
-  if (val && val.then) {
-    return Bluebird.resolve(val);
+function bluebirdUnwrapRejection<T>(rejection: T): T;
+function bluebirdUnwrapRejection(rejection: any): any {
+  if (rejection && rejection.isOperational && rejection.cause) {
+    throw rejection.cause;
   }
-  return promise;
+  throw rejection;
 }
 
 class TimeoutError extends Error {
@@ -35,29 +31,58 @@ class TimeoutError extends Error {
   }
 }
 
-/**
- * This is a helper class used to detect when a middleware wants to send a response to an RPC message.
- * The way we do this with promises is by throwing an instance of this class and the catching it.
- */
-class MiddlewareResponse extends Error {
-  response: any;
-  constructor(response) {
-    super();
-    this.response = response;
-  }
+interface RPCRequestEnvelope<TData> {
+  d: TData;
+  expiration: number;
+  returnQueue: string;
 }
 
 let HAS_WARNED_ABOUT_V8_BREAKING_CHANGE = false;
 
-class Rabbitr extends EventEmitter {
+class Rabbitr {
+  private eventListeners: {
+    [eventName: string]: Rabbitr.IEventListener<any>;
+  } = {};
+
+  on<Data>(eventName: string, listener: Rabbitr.IEventListener<Data>): void;
+  on(eventName: string, listener: Rabbitr.IEventListener<any>): void;
+
+  on(eventName: string, listener: Rabbitr.IEventListener<any>): void {
+    if (this.eventListeners[eventName]) {
+      throw new Error(`Adding multiple listeners to the same event is not supported.`);
+    }
+
+    if (typeof listener !== 'function') {
+      throw new Error(`Invalid argument passed to Rabbitr#on: ${typeof listener}. Expected a function.`);
+    }
+
+    this.eventListeners[eventName] = listener;
+  }
+
+  off(eventName: string) {
+    if (!this.eventListeners[eventName]) {
+      throw new Error(`Attempted to remove a non-existent event listener: ${eventName}`);
+    }
+    this.eventListeners[eventName] = null;
+  }
+
+  private trigger(eventName: string, message: Rabbitr.IMessage<any>) {
+    if (!this.eventListeners[eventName]) {
+      throw new Error(`Triggering an event without a listener: ${eventName}`);
+    }
+    return this.eventListeners[eventName](message);
+  }
+
+  private removeAllListeners(): void {
+    this.eventListeners = {};
+  }
+
   opts: Rabbitr.IOptions;
 
   /** @deprecated */
   protected ready = false;
   /** @deprecated */
   protected connected = false;
-
-  middleware = new Array<Rabbitr.Middleware>();
 
   protected connection: amqplib.Connection;
 
@@ -72,8 +97,6 @@ class Rabbitr extends EventEmitter {
   private debugChannelsWhitelist: string[] | void;
 
   constructor(opts: Rabbitr.IOptions) {
-    super();
-
     if (!HAS_WARNED_ABOUT_V8_BREAKING_CHANGE) {
       console.warn('Rabbitr has a major breaking change in version 8 - rpcListener queues are no longer durable. You will need to remove all rpcListener queues from RabbitMQ during deployment.')
       HAS_WARNED_ABOUT_V8_BREAKING_CHANGE = true;
@@ -89,8 +112,8 @@ class Rabbitr extends EventEmitter {
       defaultRPCExpiry: DEFAULT_RPC_EXPIRY,
     }, opts);
     this.opts.connectionOpts = objectAssign({
-      heartbeat: 1
-    }, opts && opts.connectionOpts || {});
+      heartbeat: 1,
+    }, opts && opts.connectionOpts);
 
     // istanbul ignore next
     if (!this.opts.url) {
@@ -127,7 +150,7 @@ class Rabbitr extends EventEmitter {
     if (this.isShuttingDown) {
       if (this.pendingMessagesCount) {
         // wait
-        this.log(`we have ${yellow(this.pendingMessagesCount + '')} pending messsges`);
+        this.log(`we have ${yellow(`${this.pendingMessagesCount}`)} pending messsges`);
       } else {
         return this.destroy();
       }
@@ -174,28 +197,28 @@ class Rabbitr extends EventEmitter {
           this.connected = true;
 
           this.log('ready');
-          return maybeFromCallback<void>(this.opts.setup || (() => Bluebird.resolve()))
+          return Bluebird.resolve(this.opts.setup ? this.opts.setup() : null)
             .then(() => {
               this.ready = true;
               return conn;
             });
         });
       });
-    }).catch(function(error) {
+    }).catch(error => {
       // istanbul ignore next
       process.nextTick(() => { throw error; });
     });
   }
 
   // istanbul ignore next
-  public whenReady(callback?: () => void): Bluebird<void> {
-    return this.connectionPromise.then<void>(() => void 0).asCallback(callback);
+  public whenReady(): Bluebird<void> {
+    return this.connectionPromise.then<void>(() => void 0);
   }
 
   private _formatName(name: string): string {
     // istanbul ignore next
     if (this.opts.queuePrefix) {
-      name = this.opts.queuePrefix + '.' + name;
+      return `${this.opts.queuePrefix}.${name}`;
     }
 
     return name;
@@ -203,7 +226,7 @@ class Rabbitr extends EventEmitter {
 
   private destroyPromise: Bluebird<void>;
   /** method to destroy anything for this instance of rabbitr */
-  destroy(cb?: Rabbitr.ErrorCallback): Bluebird<void> {
+  destroy(): Bluebird<void> {
     if (!this.destroyPromise) {
       this.log(`${red('destroying')}`);
       this.destroyPromise = Bluebird.each(this._openChannels, channel =>
@@ -233,20 +256,20 @@ class Rabbitr extends EventEmitter {
         );
       });
     }
-    return this.destroyPromise.asCallback(cb);
+    return this.destroyPromise;
   }
 
   // standard pub/sub stuff
 
-  send(topic: string, data: any, cb?: Rabbitr.ErrorCallback, opts?: Rabbitr.ISendOptions): Bluebird<void>;
-  send<TInput>(topic: string, data: TInput, cb?: Rabbitr.ErrorCallback, opts?: Rabbitr.ISendOptions): Bluebird<void>;
+  send(topic: string, data: any, opts?: Rabbitr.ISendOptions): Bluebird<void>;
+  send<TInput>(topic: string, data: TInput, opts?: Rabbitr.ISendOptions): Bluebird<void>;
 
-  send<TInput>(topic: string, data: TInput, cb?: Rabbitr.ErrorCallback, opts?: Rabbitr.ISendOptions): Bluebird<void> {
+  send<TInput>(topic: string, data: TInput, opts?: Rabbitr.ISendOptions): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.send(topic, data, cb, opts)
+        this.send(topic, data, opts)
       );
     }
 
@@ -255,35 +278,27 @@ class Rabbitr extends EventEmitter {
     return Bluebird.fromCallback(callback =>
       this._publishChannel.assertExchange(this._formatName(topic), 'topic', {}, callback)
     ).then(() => {
-      this._publishChannel.publish(this._formatName(topic), '*', new Buffer(stringify(data)), {
-        contentType: 'application/json',
-      });
-    }).asCallback(cb);
+      this._publishChannel.publish(
+        this._formatName(topic),
+        '*',
+        new Buffer(stringify(data)),
+        {
+          contentType: 'application/json',
+          headers: opts && opts.headers,
+        }
+      );
+    });
   }
 
-  on(topic: string, cb: (data: Rabbitr.IMessage<any>) => void): this;
-  on<TData>(topic: string, cb: (data: Rabbitr.IMessage<TData>) => void): this;
+  subscribe(topic: string, opts?: Rabbitr.ISubscribeOptions): Bluebird<void>;
+  subscribe<TMessage>(topic: string, opts?: Rabbitr.ISubscribeOptions): Bluebird<void>;
 
-  /** @private */
-  on(topic: string, cb: (data: Rabbitr.IEnvelopedMessage<any>) => void): this;
-
-  subscribe(topic: string, cb?: Rabbitr.Callback<any>): Bluebird<void>;
-  subscribe(topic: string, opts?: Rabbitr.ISubscribeOptions, cb?: Rabbitr.Callback<any>): Bluebird<void>;
-  subscribe<TMessage>(topic: string, cb?: Rabbitr.Callback<TMessage>): Bluebird<void>;
-  subscribe<TMessage>(topic: string, opts: Rabbitr.ISubscribeOptions, cb?: Rabbitr.Callback<TMessage>): Bluebird<void>;
-
-  subscribe<TMessage>(topic: string, opts?: Rabbitr.ISubscribeOptions, cb?: Rabbitr.ErrorCallback): Bluebird<void> {
-    // istanbul ignore next
-    if (typeof opts === 'function') {
-      cb = <any>opts;
-      opts = null;
-    }
-
+  subscribe<TMessage>(topic: string, opts?: Rabbitr.ISubscribeOptions): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.subscribe(topic, opts, cb)
+        this.subscribe(topic, opts)
       );
     }
 
@@ -300,11 +315,7 @@ class Rabbitr extends EventEmitter {
 
     return Bluebird.fromCallback<amqplib.Channel>(callback =>
       this.connection.createChannel(callback)
-    ).catch(error => {
-      // istanbul ignore next
-      this.emit('error', error);
-      throw error;
-    }).then(channel => {
+    ).then(channel => {
       this._openChannels.push(channel);
 
       return Bluebird.fromCallback(callback =>
@@ -329,27 +340,25 @@ class Rabbitr extends EventEmitter {
 
           this.log(`got a new message on ${cyan(topic)}`, data);
 
-          const messageAcknowledgement = new Bluebird((ack: () => void, reject) => {
+          const messageAcknowledgement = Bluebird.try(() => {
             const message: Rabbitr.IMessage<TMessage> = {
               send: this.send.bind(this),
               rpcExec: this.rpcExec.bind(this),
+
               topic,
               data,
               channel,
-              ack,
-              reject,
+              isRPC: false,
+              headers: msg.properties.headers,
             };
 
             if (options && options.skipMiddleware) {
-              this.emit(topic, message);
-              return null;
+              return this.trigger(topic, message);
             }
 
-            this._runMiddleware(message).then(() => {
-              // TODO - how to handle common error function thing for middleware?
-              this.emit(topic, message);
-              return null;
-            });
+            return this.useMiddleware(message, () =>
+              this.trigger(topic, message)
+            );
           }).then(
             // acknowledged
             () => {
@@ -375,12 +384,6 @@ class Rabbitr extends EventEmitter {
                 throw new TimeoutError({isRpc: false, topic});
               }),
           ]).catch(TimeoutError, error => {
-            this.emit('warning', {
-              type: 'ack.timeout',
-              queue: topic,
-              message: data,
-            });
-
             if (this.opts.autoAckOnTimeout === 'acknowledge') {
               channel.ack(msg);
             } else if (this.opts.autoAckOnTimeout === 'reject') {
@@ -395,21 +398,17 @@ class Rabbitr extends EventEmitter {
 
         return Bluebird.fromCallback(callback =>
           channel.consume(this._formatName(topic), processMessage, {}, callback)
-        ).catch(error => {
-          // istanbul ignore next
-          this.emit('error', error);
-          throw error;
-        });
-      }).asCallback(cb);
+        );
+      });
     });
   }
 
-  bindExchangeToQueue(exchange: string, queue: string, cb?: Rabbitr.ErrorCallback): Bluebird<void> {
+  bindExchangeToQueue(exchange: string, queue: string): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.bindExchangeToQueue(exchange, queue, cb)
+        this.bindExchangeToQueue(exchange, queue)
       );
     }
 
@@ -417,24 +416,16 @@ class Rabbitr extends EventEmitter {
 
     return Bluebird.fromCallback<amqplib.Channel>(callback =>
       this.connection.createChannel(callback)
-    ).catch(error => {
-      // istanbul ignore next
-      this.emit('error', error);
-      throw error;
-    }).then(channel => {
+    ).then(channel => {
       channel.assertQueue(this._formatName(queue));
       channel.assertExchange(this._formatName(exchange), 'topic');
 
       return Bluebird.fromCallback(callback =>
         channel.bindQueue(this._formatName(queue), this._formatName(exchange), '*', {}, callback)
-      ).catch(error => {
-        // istanbul ignore next
-        this.emit('error', error);
-        throw error;
-      }).then(ok => {
+      ).then(ok => {
         return Bluebird.fromCallback(callback => channel.close(callback));
       });
-    }).asCallback(cb);
+    });
   };
 
   // timed queue stuff
@@ -442,12 +433,12 @@ class Rabbitr extends EventEmitter {
     return `dlq.${topic}.${uniqueID}`;
   }
 
-  setTimer<TData>(topic: string, uniqueID: string, data: TData, ttl: number, cb?: Rabbitr.ErrorCallback): Bluebird<void> {
+  setTimer<TData>(topic: string, uniqueID: string, data: TData, ttl: number, opts?: Rabbitr.ISetTimerOptions): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.setTimer(topic, uniqueID, data, ttl, cb)
+        this.setTimer(topic, uniqueID, data, ttl)
       );
     }
 
@@ -464,25 +455,21 @@ class Rabbitr extends EventEmitter {
         },
         expires: (ttl + 1000)
       }, callback)
-    ).catch(error => {
-      // istanbul ignore next
-      this.emit('error', error);
-      throw error;
-    }).then(() => {
+    ).then(() => {
       this._timerChannel.sendToQueue(this._formatName(timerQueue), new Buffer(stringify(data)), {
         contentType: 'application/json',
-        // TODO - should we do anything with this?
         expiration: `${ttl}`,
+        headers: opts && opts.headers,
       });
-    }).asCallback(cb);
+    });
   }
 
-  clearTimer(topic: string, uniqueID: string, cb?: Rabbitr.ErrorCallback): Bluebird<void> {
+  clearTimer(topic: string, uniqueID: string): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.clearTimer(topic, uniqueID, cb)
+        this.clearTimer(topic, uniqueID)
       );
     }
 
@@ -492,7 +479,7 @@ class Rabbitr extends EventEmitter {
 
     return Bluebird.fromCallback(callback =>
       this._timerChannel.deleteQueue(timerQueue, {}, callback)
-    ).asCallback(cb);
+    );
   }
 
   // rpc stuff
@@ -523,31 +510,22 @@ class Rabbitr extends EventEmitter {
     });
   }
 
-  rpcExec(topic: string, data: any, cb?: Rabbitr.Callback<any>): Bluebird<any>;
-  rpcExec(topic: string, data: any, opts: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<any>): Bluebird<any>;
-  rpcExec<TInput, TOutput>(topic: string, data: TInput, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput>;
-  rpcExec<TInput, TOutput>(topic: string, data: TInput, opts: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput>;
+  rpcExec(topic: string, data: any, opts?: Rabbitr.IRpcExecOptions): Bluebird<any>;
+  rpcExec<TInput, TOutput>(topic: string, data: TInput, opts?: Rabbitr.IRpcExecOptions): Bluebird<TOutput>;
 
-  rpcExec<TInput, TOutput>(topic: string, data: TInput, opts?: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput> {
+  rpcExec<TInput, TOutput>(topic: string, data: TInput, opts?: Rabbitr.IRpcExecOptions): Bluebird<TOutput> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.rpcExec<TInput, TOutput>(topic, data, opts, cb)
+        this.rpcExec<TInput, TOutput>(topic, data, opts)
       );
-    }
-
-    // istanbul ignore next
-    if ('function' === typeof opts) {
-      // shift arguments
-      cb = <Rabbitr.Callback<TOutput>>opts;
-      opts = null;
     }
 
     // this will send the data down the topic and then open up a unique return queue
     const rpcQueue = this._rpcQueueName(topic);
 
-    const unique = v4() + '_' + ((Math.round(new Date().getTime() / 1000) + '').substr(5));
+    const unique = `${v4()}_${((`${Math.round(new Date().getTime() / 1000)}`).substr(5))}`;
     const returnQueueName = `${rpcQueue}.return.${unique}`;
 
     const now = new Date().getTime();
@@ -577,23 +555,24 @@ class Rabbitr extends EventEmitter {
 
       return Bluebird.fromCallback(callback =>
         channel.consume(replyQueue, gotReply, {noAck: true}, callback)
-      ).catch(error => {
-        // istanbul ignore next
-        this.emit('error', error);
-        throw error;
-      }).then<TOutput>(() => {
+      ).then<TOutput>(() => {
         // send the request now
-        const request = {
+        const request: RPCRequestEnvelope<TInput> = {
           d: data,
           returnQueue: this._formatName(returnQueueName),
           expiration: now + timeoutMS,
         };
 
         this.log('sending rpc request');
-        this._publishChannel.sendToQueue(this._formatName(rpcQueue), new Buffer(stringify(request)), {
-          contentType: 'application/json',
-          expiration: `${timeoutMS}`,
-        });
+        this._publishChannel.sendToQueue(
+          this._formatName(rpcQueue),
+          new Buffer(stringify(request)),
+          {
+            contentType: 'application/json',
+            expiration: `${timeoutMS}`,
+            headers: opts && opts.headers,
+          }
+        );
 
         return Bluebird.race<amqplib.Message>([
           // set a timeout
@@ -619,28 +598,25 @@ class Rabbitr extends EventEmitter {
               }
             }
 
+            // FIXME - find a way to return msg.properties.headers as well
+
             return response;
-          }, error => {
-            // TODO - investigate why bluebird wraps the error here with an object
-            if (error && error.isOperational && error.cause) {
-              throw error.cause;
-            }
-            throw error;
-          }
+          },
+          // TODO - investigate why bluebird wraps the error here
+          //   with an object
+          bluebirdUnwrapRejection
         ).catch(TimeoutError, error => {
           this.log(`request timeout firing for ${rpcQueue} to ${returnQueueName}`);
           throw error;
         });
       });
-    }).asCallback(cb);
+    });
   }
 
-  rpcListener(topic: string, executor: Rabbitr.IRpcListenerExecutor<any, any>, callback?: Rabbitr.ErrorCallback): Bluebird<void>;
-  rpcListener(topic: string, opts: Rabbitr.IRpcListenerOptions<any, any>, executor: Rabbitr.IRpcListenerExecutor<any, any>, callback?: Rabbitr.ErrorCallback): Bluebird<void>;
-  rpcListener<TInput, TOutput>(topic: string, executor: Rabbitr.IRpcListenerExecutor<TInput, TOutput>, callback?: Rabbitr.ErrorCallback): Bluebird<void>;
-  rpcListener<TInput, TOutput>(topic: string, opts: Rabbitr.IRpcListenerOptions<TInput, TOutput>, executor: Rabbitr.IRpcListenerExecutor<TInput, TOutput>, callback?: Rabbitr.ErrorCallback): Bluebird<void>;
+  rpcListener(topic: string, opts: Rabbitr.IRpcListenerOptions<any, any>, executor: Rabbitr.IRpcListenerExecutor<any, any>): Bluebird<void>;
+  rpcListener<TInput, TOutput>(topic: string, opts: Rabbitr.IRpcListenerOptions<TInput, TOutput>, executor: Rabbitr.IRpcListenerExecutor<TInput, TOutput>): Bluebird<void>;
 
-  rpcListener<TInput, TOutput>(topic: string, opts: Rabbitr.IRpcListenerOptions<TInput, TOutput>, executor?, callback?: Rabbitr.ErrorCallback): Bluebird<void> {
+  rpcListener<TInput, TOutput>(topic: string, opts: Rabbitr.IRpcListenerOptions<TInput, TOutput>, executor?): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
@@ -649,86 +625,63 @@ class Rabbitr extends EventEmitter {
       );
     }
 
-    // istanbul ignore next
-    if ('function' === typeof opts) {
-      // shift arguments
-      callback = executor as any;
-      executor = opts as any;
-      opts = {};
-    }
-
     var rpcQueue = this._rpcQueueName(topic);
 
     this.log(`has rpcListener for ${topic}`);
 
-    this.on(rpcQueue, (envelope: Rabbitr.IEnvelopedMessage<TInput>) => {
-      const dataEnvelope = envelope.data;
+    this.on(rpcQueue, (envelopedMessage: Rabbitr.IMessage<RPCRequestEnvelope<TInput>>): Bluebird<void> => {
+      const envelope = envelopedMessage.data;
 
+      // discard expired messages
       const now = new Date().getTime();
-
-      if (now > dataEnvelope.expiration) {
-        envelope.ack();
-        return;
+      if (now > envelope.expiration) {
+        return Bluebird.resolve();
       }
 
-      const message: Rabbitr.IMessage<TInput> = <Rabbitr.IEnvelopedMessage<TInput> & Rabbitr.IMessage<TInput>>envelope;
-      message.data = dataEnvelope.d;
-
-      // support for older clients - is this needed?
-      message.queue = {
-        shift: message.ack,
-      };
-
-      // TODO - how to handle common error function thing for middleware?
-      this._runMiddleware(message).then(() => {
-        return Bluebird.reduce( // run middleware
-            opts.middleware || [], (memo: void, middlewareFunc: Function) => {
-            return Bluebird.fromCallback<void>(callback => middlewareFunc(message, callback, (err, response) => {
-              callback(err || new MiddlewareResponse(response));
-            }));
-          }, null
-        )
-          .then<TOutput>(() => // then executor
-            maybeFromCallback(executor.bind(null, message))
-          )
-          .catch(error => {
-            // TODO - investigate why bluebird wraps the error here with an object
-            if (error && error.isOperational && error.cause) {
-              throw error.cause;
-            }
-            throw error;
-          })
-          .catch(MiddlewareResponse, ({response}) => { // catch middleware responses
-            return response
-          })
-          .then<any>( // sanitize errors
-            response => {
-              this.log(`${yellow('rpcListener')} responding to topic ${cyan(topic)} with`, response);
-              return {response};
-            }, error => {
-              this.log(`${yellow('rpcListener')} on topic ${cyan(topic)} ${red('hit error')}`, error);
-
-              var isError = error instanceof Error;
-              var errJSON = isError ?
-                stringifyError(error) :
-                stringify(error);
-
-              return {
-                error: errJSON,
-                isError,
-              };
-            }
-          )
-          .then(data => { // send the response
-            // ack here - this will get ignored if the executor has acked or nacked already anyway
-            message.ack();
-
-            // doesn't need wrapping in this.formatName as the rpcExec function already formats the return queue name as required
-            this._publishChannel.sendToQueue(dataEnvelope.returnQueue, new Buffer(stringify(data)), {
-              contentType: 'application/json',
-            });
-          }); // TODO - log uncaught errors at this stage? bluebird will do it anyway.
+      const message: Rabbitr.IMessage<TInput> = objectAssign({}, envelopedMessage, {
+        data: envelope.d,
+        isRPC: true,
+        responseHeaders: {},
       });
+
+      return this.useMiddleware(message, executor.bind(null, message))
+        .catch(
+          // TODO - investigate why bluebird wraps the error here
+          //   with an object
+          bluebirdUnwrapRejection
+        )
+        .then<any>( // sanitize errors
+          response => {
+            this.log(`${yellow('rpcListener')} responding to topic ${cyan(topic)} with`, response);
+            return {response};
+          }, error => {
+            this.log(`${yellow('rpcListener')} on topic ${cyan(topic)} ${red('hit error')}`, error);
+
+            var isError = error instanceof Error;
+            var errJSON = isError ?
+              stringifyError(error) :
+              stringify(error);
+
+            return {
+              error: errJSON,
+              isError,
+            };
+          }
+        )
+        .then(data => { // send the response
+          this._publishChannel.sendToQueue(
+            // doesn't need wrapping in this.formatName as the rpcExec function
+            //   already formats the return queue name as required
+            envelope.returnQueue,
+            new Buffer(stringify(data)),
+            {
+              contentType: 'application/json',
+              headers: message.responseHeaders,
+            }
+          );
+        });
+        // TODO - log uncaught errors at this stage?
+        //   bluebird will do it anyway.
     });
 
     return this.subscribe(rpcQueue,
@@ -736,18 +689,24 @@ class Rabbitr extends EventEmitter {
         skipMiddleware: true,
         durable: false,
       })
-    ).asCallback(callback);
+    );
   }
 
-  // message middleware support
-  use(middlewareFunc: Rabbitr.Middleware) {
-    this.middleware.push(middlewareFunc);
+  // #region middleware
+  private middlewareFn = new Array<Rabbitr.Middleware>();
+
+  middleware(fn: Rabbitr.Middleware): void {
+    this.middlewareFn.push(fn);
   }
-  private _runMiddleware(message: Rabbitr.IMessage<any>): Bluebird<void> {
-    return Bluebird.reduce(this.middleware, (memo: void, middlewareFunc) => {
-      return Bluebird.fromCallback<void>(callback => middlewareFunc(message, callback));
-    }, null);
+
+  private useMiddleware<T>(message: Rabbitr.IMessage<any>, next: () => T | PromiseLike<T>): Bluebird<T> {
+    return Bluebird.try(
+      this.middlewareFn.reduce<typeof next>(function(next, middlewareFn) {
+        return middlewareFn.bind(null, message, next) as typeof next;
+      }, next)
+    );
   }
+  // #endregion middleware
 };
 
 declare module Rabbitr {
@@ -760,7 +719,7 @@ declare module Rabbitr {
     queuePrefix?: string;
 
     /** called once the connection is ready but before anything is bound (allows for ORM setup etc) */
-    setup?: ((done: Rabbitr.ErrorCallback) => void) | (() => PromiseLike<void>);
+    setup?: () => void | PromiseLike<void>;
     connectionOpts?: {
       heartbeat?: boolean;
     };
@@ -769,25 +728,21 @@ declare module Rabbitr {
     defaultRPCExpiry?: number;
   }
 
-  export interface ErrorCallback {
-    (err: Error): void;
-  }
+  export type IEventListener<TData> =
+    ((message: IMessage<TData>) => void | PromiseLike<void>);
 
-  export interface Callback<T> {
-    (err: Error): void;
-    (err: Error, data: T): void;
-  }
+  export interface Headers { [header: string]: string; }
 
   export interface IRpcExecOptions {
     timeout?: number;
+    headers?: any; // Headers;
   }
   export interface IRpcListenerOptions<TInput, TOutput> {
     middleware?: Function[];
     prefetch?: number;
   }
   export type IRpcListenerExecutor<TInput, TOutput> =
-    ((message: IMessage<TInput>) => PromiseLike<TOutput>) |
-    ((message: IMessage<TInput>, respond: Callback<TOutput>) => void);
+    ((message: IMessage<TInput>) => TOutput | PromiseLike<TOutput>);
 
   export interface ISubscribeOptions {
     prefetch?: number;
@@ -795,40 +750,37 @@ declare module Rabbitr {
     durable?: boolean;
   }
   export interface ISendOptions {
+    headers?: any; // Headers;
+  }
+  export interface ISetTimerOptions {
+    headers?: any; // Headers;
   }
 
   export interface IMessage<TData> {
-    ack(): void;
-    reject(error?: Error): void;
-
     topic: string;
     channel: amqplib.Channel;
     data: TData;
+    headers: { [header: string]: string; };
 
-    send(topic: string, data: any, cb?: Rabbitr.ErrorCallback, opts?: Rabbitr.ISendOptions): Bluebird<void>;
-    send<TInput>(topic: string, data: TInput, cb?: Rabbitr.ErrorCallback, opts?: Rabbitr.ISendOptions): Bluebird<void>;
+    send(topic: string, data: any, opts?: Rabbitr.ISendOptions): Bluebird<void>;
+    send<TInput>(topic: string, data: TInput, opts?: Rabbitr.ISendOptions): Bluebird<void>;
 
-    rpcExec(topic: string, data: any, cb?: Rabbitr.Callback<any>): Bluebird<any>;
-    rpcExec(topic: string, data: any, opts: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<any>): Bluebird<any>;
-    rpcExec<TInput, TOutput>(topic: string, data: TInput, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput>;
-    rpcExec<TInput, TOutput>(topic: string, data: TInput, opts: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput>;
+    rpcExec(topic: string, data: any): Bluebird<any>;
+    rpcExec(topic: string, data: any, opts: Rabbitr.IRpcExecOptions): Bluebird<any>;
+    rpcExec<TInput, TOutput>(topic: string, data: TInput, opts?: Rabbitr.IRpcExecOptions): Bluebird<TOutput>;
 
-    queue?: {
-      shift: () => void;
+    isRPC: boolean;
+    /** only for rpc: message headers to be sent back with the response */
+    responseHeaders?: {
+      [header: string]: string;
     };
   }
 
-  export interface IEnvelopedMessage<TData> extends IMessage<any> {
-    data: {
-      d: TData,
-      expiration: number,
-      returnQueue: string,
-    };
+  export interface MiddlewareCallback {
+    (): Bluebird<any | void>;
   }
-
   export interface Middleware {
-    // TODO - better annotation
-    (message: IMessage<any>, cb: Function, next?: Function): void;
+    (fn: (message: IMessage<any>, next: MiddlewareCallback) => PromiseLike<any | void>): void;
   }
 }