+++ test/pubsub.ts
@@ -33,14 +33,14 @@ describe('rabbitr#pubsub', function() {
   );
 
   it('should receive messages on the specified queue', function(done) {
-    const exchangeName = v4() + '.pubsub_test';
-    const queueName = v4() + '.pubsub_test';
+    const exchangeName = `${v4()}.pubsub_test`;
+    const queueName = `${v4()}.pubsub_test`;
 
     const testData = {
-      testProp: 'pubsub-example-data-' + queueName
+      testProp: `pubsub-example-data-${queueName}`
     };
 
-    rabbit.subscribe(queueName)
+    rabbit.subscribe(queueName, {})
       .then(() => createdQueues.push(queueName))
       .then(() =>
         rabbit.bindExchangeToQueue(exchangeName, queueName)
@@ -48,14 +48,14 @@ describe('rabbitr#pubsub', function() {
             createdExchanges.push(exchangeName)
           )
           .then(() =>
-            rabbit.send(exchangeName, testData)
+            rabbit.send(exchangeName, testData, {})
           )
       );
 
-    rabbit.on(queueName, function(message) {
+    rabbit.on(queueName, function(message, reply) {
       Bluebird
         .try(() => {
-          message.ack();
+          reply();
 
           // here we'll assert that the data is the same- the fact we received it means the test has basically passed anyway
           expect(JSON.stringify(testData)).to.equal(JSON.stringify(message.data));
@@ -64,9 +64,39 @@ describe('rabbitr#pubsub', function() {
     });
   });
 
+  it('should receive messages on the specified queue (shifted arguments)', function(done) {
+    const exchangeName = `${v4()}.pubsub_test`;
+    const queueName = `${v4()}.pubsub_test`;
+
+    const testData = {
+      testProp: `pubsub-example-data-${queueName}`,
+    };
+
+    Bluebird.fromCallback(cb => rabbit.subscribe(queueName, cb))
+      .then(() => createdQueues.push(queueName))
+      .then(() =>
+        rabbit.bindExchangeToQueue(exchangeName, queueName)
+          .then(() =>
+            createdExchanges.push(exchangeName)
+          )
+          .then(() =>
+            Bluebird.fromCallback(cb => rabbit.send(exchangeName, testData, cb))
+          )
+      );
+
+    rabbit.on(queueName, function(message) {
+      Bluebird
+        .try(() => {
+          // here we'll assert that the data is the same- the fact we received it means the test has basically passed anyway
+          expect(JSON.stringify(testData)).to.equal(JSON.stringify(message.data));
+        })
+        .asCallback(done);
+    });
+  });
+
   it('passes Buffers', function(done) {
-    const exchangeName = v4() + '.pubsub_buf_test';
-    const queueName = v4() + '.pubsub_buf_test';
+    const exchangeName = `${v4()}.pubsub_buf_test`;
+    const queueName = `${v4()}.pubsub_buf_test`;
 
     const data = 'Hello world!';
 
@@ -85,8 +115,6 @@ describe('rabbitr#pubsub', function() {
     rabbit.on(queueName, function(message) {
       Bluebird
         .try(() => {
-          message.ack();
-
           expect(message.data).to.be.an.instanceOf(Buffer);
           expect(message.data.toString()).to.equal(data);
         })